tools/requirements.txt
data==0.4
matplotlib==3.9.1
mlflow==2.14.3
numpy==2.0.1
Pillow==9.4.0
Pillow==10.4.0
ray==2.33.0
scikit_learn==1.5.1
seaborn==0.13.2
tools==0.1.9
torch==2.1.0
